requirements.txt
pillow==6.0.0
requests==2.22.0